pyproject.toml
[project]
name = "rag-series"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.12",
    "faster-whisper>=1.1.1",
    "google-generativeai>=0.3.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=0.0.10",
    "langgraph>=0.4.7",
    "pydub>=0.25.1",
    "pypdf>=5.6.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
    "transformers>=4.52.4",
    "yt-dlp>=2025.5.22",
]
